requirements.txt
duckduckgo_search==8.0.1
openai==1.75.0
Pillow
pydantic==2.11.3
pyprojroot==0.3.0
python-dotenv==1.1.0
PyYAML==6.0.2
streamlit==1.44.1
streamlit-chat==0.1.1
